requirements.txt
mediapipe >= 0.10.0
opencv-python >= 4.8.0
pyobjc-framework-Quartz >= 10.0
